Implemented a TransformTree type. Ideas are starting to crystallize around how to organize geometry, collision checking, etc. I'm thinking there should be four main types of "entity" nodes: - Robot - Object - Assembly (a collection of objects and their transforms relative to the   assembly frame) - TransportUnit (an assembly and its transform relative to the   transport_unit frame, as well as one or more robots with their transforms to the transport_unit frame). With correct design, all path planning can be done for TransportUnits only. Or maybe we also want to plan paths for Robot nodes. The collision checking/deferred geometry computation should be organized (somehow) around this framework. At this point, I'm not entirely sure about WHERE to store information. Should there be a transform tree that exists separately from all geometry? Should everything be stored within the nodes of the same tree? Should there be three graphs? An entity graph, a schedule graph, and a transform tree? Still trying to wrap my head around the details.
A few other notes: The original untransformed geometry needs to be
stored in a location that is easily accessible and well organized. This
is also the case for the local transforms that must be defined by
Assembly and TransportUnit types. There should be a clean way of
extracting a concise, complete representation of the environment state
(transforms of all entities, as well as some indicator of which entities
are active). Also, don't forget that the base geometry for LEGO
assemblies, along with the assemblies and their transforms, are already
accessible through the LDrawParser package.

Also: the current "object" geometry (in the TaskGraphs package) can be
thought of as NULL, in the sense that nothing can collide with an object.
The robots are just disks.

All tests passing.