get_cached_geom(...) is now type stable for Ball2, Hyperrectangle, Polytopes, and Vectors of Ngons. But the polyhedron conversion to VPolytope is hanging endlessly.
I looked online and saw that the endless hanging (not actually
endless--just several minutes) is actually caused by _printing_ the
polytope--not by converting between VPolytope and HPolytope.

Update: It's not just printing. It's the conversion from HPolytope to
VPolytope. Apparently the Polyhedra backend takes a long time to
compile!

Update: In my case, is is indeed HANGING. Never finishes.