Trying to figure out how to deal with 2D and 3D geometry simultaneously in the SceneTree. The purpose is to have approximations that are projected onto the X-Y plane. Is it necessary? I'm trying to decide. With a circle, the element could be computed on the fly (project the sphere onto the XY plane. With a polygon, I could also have a 3D planar polygon that gets projected when necessary. Since Polyhedra seems to be out of the picture anyway, I might as well skip the polygon thing.
Still, it is nice to have a bounding circle around the geometry
projected to the X-Y plane.

There are two main needs to address: First, how to define simple
dispatch rules that will reliable project 3D geometry to 2D to create 2D
approximations.
Second, how to reliably transform 2D geometry with the other geometry
elements of the scene tree.

Options:
- Have GeomNode2D vs GeomNode3D to distinguish between geometry types.
- ??